% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Library 1.R
\name{ChkPkgANDInstall}
\alias{ChkPkgANDInstall}
\title{CheckPackageANDInstall}
\usage{
ChkPkgANDInstall(packageName)
}
\description{
Check whether package exists, otherwise install.
}
\examples{
ChkPkgANDInstall()
}

