% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Library_1.R
\name{describeData}
\alias{describeData}
\title{Describe Dataset}
\usage{
describeData(df)
}
\description{
This function describes the columns in the dataset. Dataframe is sent as the input
}
\examples{
describeData()
}

