% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Library 1.R
\name{import.csv}
\alias{import.csv}
\title{Import CSV}
\usage{
import.csv(filename)
}
\description{
This function allows you to import a CSV.
Default input is comma separated value, and first row as headers.
}
\examples{
import.csv()
}

