% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Library_1.R
\name{import.xls}
\alias{import.xls}
\title{Import XLS (TEMPORARILY NOT WORKING)}
\usage{
import.xls(filename)
}
\description{
This function allows you to import a XLS.
Defeault input is that data is in first sheet.
}
\examples{
import.xls()
}

